Move TransferData into new package
