+++ src/main/java/editor/gui/ccp/data/CategoryTransferData.java
@@ -1,4 +1,4 @@
-package editor.gui.ccp;
+package editor.gui.ccp.data;
 
 import java.awt.datatransfer.DataFlavor;
 import java.awt.datatransfer.Transferable;
