+++ src/main/java/editor/gui/ccp/data/DeckTransferData.java
@@ -1,4 +1,4 @@
-package editor.gui.ccp;
+package editor.gui.ccp.data;
 
 import java.awt.datatransfer.DataFlavor;
 import java.awt.datatransfer.Transferable;
