+++ src/main/java/editor/gui/editor/EditorFrame.java
@@ -80,9 +80,9 @@ import editor.gui.CardTagPanel;
 import editor.gui.MainFrame;
 import editor.gui.TableSelectionListener;
 import editor.gui.ccp.CCPItems;
-import editor.gui.ccp.CategoryTransferData;
 import editor.gui.ccp.CategoryTransferHandler;
-import editor.gui.ccp.DataFlavors;
+import editor.gui.ccp.data.CategoryTransferData;
+import editor.gui.ccp.data.DataFlavors;
 import editor.gui.ccp.handler.EditorTableTransferHandler;
 import editor.gui.ccp.handler.EditorTransferHandler;
 import editor.gui.display.CardImagePanel;
